Re-organises project modules
