+++ src/cli/logging.rs
@@ -1,9 +1,12 @@
 // Copyright 2024 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
+use human_panic::setup_panic;
 use log::LevelFilter;
 
-pub fn setup(verbose: bool) {
+pub fn setup_logging(verbose: bool) {
+    setup_panic!();
+
     let filter = if verbose { LevelFilter::Debug } else { LevelFilter::Info };
 
     env_logger::Builder::new()
